# Some performance tuning for sorting