--- a/modules/packages/Sort.chpl
+++ b/modules/packages/Sort.chpl
@@ -487,16 +487,12 @@ proc sort(ref Data: [?Dom] ?eltType, comparator:?rec=defaultComparator,
       // TODO: use a sample sort if the input does not have enough
       // randomness, according to some heuristic
 
-      // TODO: uncomment the below when implementation is ready
-      /*
-      var quickSortSize=50_000;
-      // quick sort performs better for middle problem sizes
-      // (too large for insertion sort etc, but not big enough
-      //  to overcome the larger constant overhead of radix sort)
-      if Data.domain.size < quickSortSize {
-        QuickSort.quickSort(Data, comparator=comparator);
+      var simplerSortSize=50_000;
+      if Data.domain.size < simplerSortSize {
+        // TODO: use quicksort instead in these small cases
+        MSBRadixSort.msbRadixSort(Data, comparator=comparator);
         return;
-      }*/
+      }
 
       if inPlaceAlgorithm {
         // use an in-place algorithm