This PR makes some performance adjustments to Sort.chpl:

 * adjust `sort` to use the older in-place radix sorter for problems too small for TwoArrayRadixSort before creating the Scratch array (because it will not be needed for small subproblems). My experiments indicate using QuickSort in these cases is even better but we aren't quite ready to do that yet.
 * added a skip-ahead mechanism for the two-array radix sort for when multiple key bytes are the same for all elements. To do this, made the `startbit` formal to `bucketize` use `inout` intent; so it might change the start bit being considered during the counting process (if all the data starts with the same pattern)
 * it adds throttling for the case of sorting within a nested parallel region

### Performance Impact

The tables below contain data generated by `sort-performance-explorer.chpl` with and without this PR to see the performance impact. That benchmark supports a bunch of different input patterns, so the tables have different rows for these.

This table shows the change in performance due to the skip-ahead. It's using a problem size of 1GiB of `int`s. Cells are speeds in MiB/s:

|            | before this PR | this PR | note |
|------------|----------------|---------|------|
| all zeros  | 1700.22        | 5329.89 | ~3x faster |
| random     | 4195.17        | 4199.01 | equivalent |
| top byte   | 7185.71        | 7186.77 | equivalent |
| middle byte| 3491.47        | 5511.66 | ~1.5x faster |
| bottom byte| 1702.92        | 4027.71 | ~2x faster |
| powers of 2| 2265.12        | 2203.17 | 3% slower |
| 1..n       | 2042.67        | 3845.47 | ~2x faster |
| 1..n by -1 | 2060.28        | 3845.95 | ~2x faster |

The next table shows the change in performance from using msbRadixSort for small problem sizes instead of the two-array sorter. Here the problem size is 512 bytes of `int`s. Cells are speeds in MiB/s:

|            | before this PR | this PR |
|------------|----------------|---------|
| all zeros  | 6.60572        | 2209.42 |
| random     | 6.56019        | 1493.21 |
| top byte   | 6.58691        | 1530.66 |
| middle byte| 6.5975         | 1470.73 |
| bottom byte| 6.69438        | 1516.4  |
| powers of 2| 6.62409        | 1570.04 |
| 1..n       | 6.59225        | 2219.46 |
| 1..n by -1 | 6.65324        | 1856.58 |

### Future Work

 * Use QuickSort instead of msbRadixSort for small problems since it has lower constant overhead & seems to be faster at these small sizes
 * Add code to do first-touch for the Scratch array in the distributed sorts, which should be relevant with `CHPL_COMM=ugni`
 * Explore using twoArraySample for large arrays with highly skewed distributions. e.g. `./sort-performance-explorer --inputDataScheme=5` shows that it's on the order of 3x faster than the radix sorter in this case. In contrast, the radix sorter is about 50% faster, in my experiments, with randomly distributed data.

### Testing and Review

Reviewed by @vasslitvinov - thanks!

- [x] full comm=none testing
- [x] full comm=gasnet testing